refactor/ST-126 move some metrics out of framework
