+++ pp/event/share.go
@@ -5,15 +5,14 @@ import (
 	"sync"
 	"time"
 
-	"github.com/stratosnet/sds/framework/crypto"
-	"github.com/stratosnet/sds/framework/metrics"
-
 	"github.com/stratosnet/sds/framework/core"
+	"github.com/stratosnet/sds/framework/crypto"
 	"github.com/stratosnet/sds/framework/msg/header"
 	fwutils "github.com/stratosnet/sds/framework/utils"
 	"github.com/stratosnet/sds/pp"
 	"github.com/stratosnet/sds/pp/api/rpc"
 	"github.com/stratosnet/sds/pp/file"
+	"github.com/stratosnet/sds/pp/metrics"
 	"github.com/stratosnet/sds/pp/p2pserver"
 	"github.com/stratosnet/sds/pp/requests"
 	"github.com/stratosnet/sds/pp/setting"
