+++ pp/task/transfer_task.go
@@ -5,9 +5,9 @@ import (
 	"time"
 
 	"github.com/pkg/errors"
+	"github.com/stratosnet/sds/pp/metrics"
 
 	"github.com/stratosnet/sds/framework/crypto"
-	"github.com/stratosnet/sds/framework/metrics"
 	"github.com/stratosnet/sds/framework/utils"
 	"github.com/stratosnet/sds/pp/file"
 	"github.com/stratosnet/sds/sds-msg/protos"
