+++ pp/task/upload_task.go
@@ -7,9 +7,9 @@ import (
 
 	"github.com/alex023/clock"
 	"github.com/pkg/errors"
+	"github.com/stratosnet/sds/pp/metrics"
 
 	"github.com/stratosnet/sds/framework/crypto"
-	"github.com/stratosnet/sds/framework/metrics"
 	"github.com/stratosnet/sds/framework/utils"
 	"github.com/stratosnet/sds/pp"
 	"github.com/stratosnet/sds/pp/file"
